Fix Windows build (#56)
* cmake: Don't use Clang/GNU command line options with MSVC

* changelog: Update changelog

* cmake: Define NOMINMAX on Windows

* cmake: Use CXX_COMPILER_ID instead of COMPILE_LANG_AND_ID

* core: Add export symbol

* [core] Remove a warning

* [core] Don't add extern with dllexport

* core: Split explicit template instantiation for constraints

* [core] Fix missing definition

* [core] Update constraints.hpp

* [core] Add missing includes

* [python] Add windows_dll_manager module to manage add_dll_directory on Windows

* [python] Fix __init__.py

* [cmake] Build Windows pyd in the right directory

* [python] Remove pinocchio reference

* [cmake] Fix path used to launch generate_stubs

* [core] Use explicit instantiation macro from jrl-cmakemodules

* cmake: Generate binaries in bin and lib

* [core] Use new define name

* [cmake] Update jrl-cmakemodules

* [changelog] Fix changelog

* [python] Add support to other site-package structure

* [ci] Update MacOS/Linux recipe

* [ci] Add Windows

* [core] Fix pre-commit

* [cmake] Use standard jrl-cmakemodules

* [ci] Fix Windows workflow

* [core] Fix some warnings

* [ci] Try to publish on Windows

* [core] Fix clang error

* [ci] Add windows to publish workflow

* [ci] Try to publish on Windows

* [core] Update LogRecord iter and al_iter instead of casting

* [cmake] Update jrl-cmakemodules to remove a warnings under ClangCl

* [ci] Add pkg-config as Windows dependency

* [cmake] Update jrl-cmakemodules to try to fix template instantiation on Windows

* [core] Use new DLLAPI macro

* [cmake] Update jrl-cmakemodules

* [core] Add dllexport on template definition

* [core] Forget some definition

* [core] Fix another missing definition

* [core] Manage MSVC error on template instantiation

* [python] Fix dll relative directory

* [ci] Test publish script on non release version

* [cmake] Use standard jrl-cmakemodules

* [changelog] Add changelog entry

+++ include/proxsuite-nlp/linalg/bunchkaufman.hpp
@@ -142,11 +142,15 @@ ComputationInfo bunch_kaufman_in_place_unblocked(MatrixType &a,
       }
     }
 
+    // pivots matrix store int by default
+    // Like Eigen, we assume that we will never
+    // have a matrix that will have more rows/columns
+    // than int can old
     if (k_step == 1) {
-      pivots[k] = kp;
+      pivots[k] = static_cast<int>(kp);
     } else {
-      pivots[k] = -1 - kp;
-      pivots[k + 1] = -1 - kp;
+      pivots[k] = static_cast<int>(-1 - kp);
+      pivots[k + 1] = static_cast<int>(-1 - kp);
     }
 
     k += k_step;
@@ -287,11 +291,15 @@ bunch_kaufman_in_place_one_block(MatrixType &a, WType &w, IndicesType &pivots,
       }
     }
 
+    // pivots matrix store int by default
+    // Like Eigen, we assume that we will never
+    // have a matrix that will have more rows/columns
+    // than int can old
     if (k_step == 1) {
-      pivots[k] = kp;
+      pivots[k] = static_cast<int>(kp);
     } else {
-      pivots[k] = -1 - kp;
-      pivots[k + 1] = -1 - kp;
+      pivots[k] = static_cast<int>(-1 - kp);
+      pivots[k + 1] = static_cast<int>(-1 - kp);
     }
 
     k += k_step;
@@ -360,10 +368,14 @@ ComputationInfo bunch_kaufman_in_place(MatrixType &a, VecType &subdiag,
 
     for (Index j = k; j < k + kb; ++j) {
       auto &p = pivots.coeffRef(j);
+      // pivots matrix store int by default
+      // Like Eigen, we assume that we will never
+      // have a matrix that will have more rows/columns
+      // than int can old
       if (p >= 0) {
-        p += k;
+        p += static_cast<int>(k);
       } else {
-        p -= k;
+        p -= static_cast<int>(k);
       }
     }
